requirements.txt
Flask==2.3.3
python-pptx==0.6.21
requests==2.31.0
google-generativeai==0.3.2
openai>=1.0.0
Pillow==10.0.1
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
markupsafe==2.1.3
pywin32==304
shutil
pptxtopdf